Added most of autorally_chassis stuff. Program is tested working within launch system on an autorally_chassis
