+++ autorally_core/CMakeLists.txt
@@ -32,10 +32,10 @@ set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} ${BUILD_FLAGS}")
 include_directories(SYSTEM ${catkin_INCLUDE_DIRS} ${Boost_INCLUDE_DIR})
 include_directories(include)
 
-add_subdirectory(src/arduino)
+#add_subdirectory(src/arduino)
 add_subdirectory(src/Diagnostics)
 add_subdirectory(src/gps)
-add_subdirectory(src/ocs)
+#add_subdirectory(src/ocs)
 add_subdirectory(src/RingBuffer)
 add_subdirectory(src/RunStop)
 add_subdirectory(src/SafeSpeed)
@@ -43,7 +43,7 @@ add_subdirectory(src/SerialSensorInterface)
 add_subdirectory(src/servoInterface)
 add_subdirectory(src/xbee)
 add_subdirectory(src/ImageRepublisher)
-add_subdirectory(src/StateEstimator)
+#add_subdirectory(src/StateEstimator)
 add_subdirectory(src/autorally_chassis)
 
 install(TARGETS
