+++ autorally_core/launch/autorally_chassis.launch
@@ -1,7 +1,7 @@
 <launch>
   <include file="$(find autorally_core)/launch/hardware.machine" />
 
-  <node pkg="nodelet" type="nodelet" name="autorally_chassis" args="load autorally_core/autorally_chassis autorally_core_manager" machine="autorally-master" output="screen">
+  <node pkg="nodelet" type="nodelet" name="AutoRallyChassis" args="load autorally_core/AutoRallyChassis autorally_core_manager" machine="autorally-master" output="screen">
 
     <param name="commandRate" value="75" />
     <param name="commandMaxAge" value="0.2" />
