+++ autorally_core/src/autorally_chassis/CMakeLists.txt
@@ -1,9 +1,9 @@
-add_library(autorally_chassis autorally_chassis.cpp)
-add_dependencies(autorally_chassis autorally_msgs_gencpp)
-target_link_libraries(autorally_chassis ${catkin_LIBRARIES} ${Boost_LIBRARIES} RingBuffer SerialSensorInterface Diagnostics)
+add_library(AutoRallyChassis AutoRallyChassis.cpp)
+add_dependencies(AutoRallyChassis autorally_msgs_gencpp)
+target_link_libraries(AutoRallyChassis ${catkin_LIBRARIES} ${Boost_LIBRARIES} RingBuffer SerialSensorInterface Diagnostics)
 
 install(TARGETS
-  autorally_chassis
+  AutoRallyChassis
 
   ARCHIVE DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
   LIBRARY DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
