+++ autorally_util/config/arChassisConfig_CHASSIS_NAME.yaml
@@ -1,3 +1,3 @@
-'0': {name: 'steering', center: 1500, min: 1000, max: 2000, reverse: false}
-'1': {name: 'throttle', center: 1500, min: 1000, max: 2000, reverse: false}
-'2': {name: 'frontBrake', center: 1500, min: 1500, max: 2000, reverse: false}
+'steering': {center: 1500, min: 1000, max: 2000, reverse: false}
+'throttle': {center: 1500, min: 1000, max: 2000, reverse: false}
+'frontBrake': {center: 1500, min: 1500, max: 2000, reverse: false}
